v1

+++ src/logViewer/saveCode.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const bodyParser = require("body-parser");
-const fs = require("fs");
+const fs = require("fs").promises;
 const { spawn } = require("child_process");
 const cors = require("cors");
 const http = require("http");
@@ -11,10 +11,11 @@ const app = express();
 const server = http.createServer(app);
 const io = socketIo(server);
 
-const port = 3000;
-
+const PORT = 3000;
 const PATH = "../../generatedCode/";
 const VENVPATH = "../../myenv/bin/python3";
+const FILENAME = "python_code.py";
+
 let currentProcess = null;
 let serverStartTime;
 
@@ -26,38 +27,25 @@ function emitLog(message) {
 
 function compilePython(code, action) {
   if (action === "run") {
-    // Clear previous logs only when running new code
     io.emit("clearLogs");
+    emitLog(`-- START --`);
 
-    // Spawn a new process to run the Python file
-    currentProcess = spawn(VENVPATH, ["-u", `${PATH}python_code.py`]);
-    console.log(`init: ${currentProcess.pid}`);
+    currentProcess = spawn(VENVPATH, ["-u", `${PATH}${FILENAME}`]);
+    console.log(`Process started with PID: ${currentProcess.pid}`);
 
-    // Handle the output
-    currentProcess.stdout.on("data", (data) => {
-      emitLog(`${data}`);
-    });
-
-    // Handle errors
-    currentProcess.stderr.on("data", (data) => {
-      emitLog(`Error: ${data}`);
-    });
-
-    // Handle process exit
+    currentProcess.stdout.on("data", (data) => emitLog(`${data}`));
+    currentProcess.stderr.on("data", (data) => emitLog(`Error: ${data}`));
     currentProcess.on("exit", (code) => {
-      emitLog(`Process exited with code ${code}`);
+      emitLog(`-- END --`);
       currentProcess = null;
     });
-  } else if (action === "stop") {
-    console.log(`stop: ${currentProcess.pid}`);
-    if (currentProcess) {
-      currentProcess.kill("SIGKILL");
-      //const stopTimestamp = performance.now() - serverStartTime;
-      //io.emit("stop", { stopTimestamp });
-      currentProcess = null;
-    } else {
-      emitLog("No running process to stop");
-    }
+  } else if (action === "stop" && currentProcess) {
+    console.log(`Stopping process: ${currentProcess.pid}`);
+    emitLog(`-- STOP --`);
+    currentProcess.kill("SIGKILL");
+    currentProcess = null;
+  } else {
+    emitLog("No running process to stop");
   }
 }
 
@@ -65,65 +53,58 @@ app.use(cors());
 app.use(bodyParser.json());
 app.use(express.static("."));
 
-app.post("/", (req, res) => {
-  res.json({
-    message: "ok",
-  });
-});
+app.post("/", (req, res) => res.json("Connected to viewer"));
 
-app.post("/save_code", (req, res) => {
-  const action = req.body.action;
-  const filename = PATH + "python_code.py";
+app.post("/save_code", async (req, res) => {
+  const { action, code } = req.body;
+  const filePath = `${PATH}${FILENAME}`;
 
-  if (action === "run") {
-    const code = req.body.code;
-    fs.writeFile(filename, code, (err) => {
-      if (err) {
-        console.error(err);
-        res.status(500).json({ error: "Failed to save the file" });
-      } else {
-        compilePython(code, action);
-        res.json({
-          message: `Code ${
-            action === "run" ? "saved and executed" : "stopped"
-          }`,
-        });
-      }
-    });
-  } else {
-    compilePython("", action);
+  try {
+    if (action === "run") {
+      await fs.writeFile(filePath, code);
+      compilePython(code, action);
+      res.json("Code saved and executed");
+    } else if (action === "stop") {
+      compilePython("", action);
+      res.json("Code stopped");
+    } else {
+      res.status(400).json("Invalid action");
+    }
+  } catch (error) {
+    console.error(error);
+    res.status(500).json("Server error");
   }
 });
 
-app.post("/run_code", (req, res) => {
-  const action = req.body.action;
-  const filename = PATH + "python_code.py";
+app.post("/run_code", async (req, res) => {
+  const { action } = req.body;
+  const filePath = `${PATH}${FILENAME}`;
 
-  console.log("run code request");
-
-  if (action === "run") {
-    fs.readFile(filename, "utf8", (err, code) => {
-      if (err) {
-        console.error(err);
-        res.status(500).json({ error: "Failed to read the file" });
-      } else {
-        compilePython(code, action);
-        res.json({
-          message: "Code read and executed",
-        });
-      }
-    });
-  } else if (action === "stop") {
-    compilePython("", action);
-    res.json({
-      message: "Code stopped",
-    });
-  } else {
-    res.status(400).json({ error: "Invalid action" });
+  try {
+    if (action === "run") {
+      const code = await fs.readFile(filePath, "utf8");
+      compilePython(code, action);
+      res.json("Code read and executed");
+    } else if (action === "stop") {
+      compilePython("", action);
+      res.json("Code stopped");
+    } else {
+      res.status(400).json("Invalid action");
+    }
+  } catch (error) {
+    try {
+      const code = `print("Hello World")`;
+      await fs.writeFile(filePath, code);
+      compilePython(code, action);
+      res.json("Code saved and executed");
+    } catch (error) {
+      console.error(error);
+      res.status(500).json("Server error");
+    }
   }
 });
 
-server.listen(port, () => {
+server.listen(PORT, () => {
   serverStartTime = performance.now();
-  console.log(`Server running at http://localhost:${port}`);
+  console.log(`Server running at http://localhost:${PORT}`);
 });
